game folder organized
